pyproject.toml
[project]
name = "market-analysis-bot"
version = "0.1.0"
description = "Cohere North MCP demo for energy trading market analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Demo Project"}
]

dependencies = [
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
]

[project.urls]
Repository = "https://github.com/cohere-ai/north-mcp-python-sdk"

[tool.uv.sources]
north-mcp-python-sdk = { git = "https://github.com/cohere-ai/north-mcp-python-sdk.git" }

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

viz = [
    "matplotlib>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=tools --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

